requirement.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.6.1
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
facebook-sdk==2.0.0
fake-useragent==0.1.10
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.3
parse==1.8.4
parsel==1.5.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
pyee==5.0.0
PyHamcrest==1.9.0
pymongo==3.7.1
pyOpenSSL==18.0.0
pyppeteer==0.0.19
pyquery==1.4.0
python-dateutil==2.7.3
pytz==2018.5
queuelib==1.5.0
requests==2.19.1
requests-html==0.9.0
Scrapy==1.5.1
selenium==3.13.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
websockets==6.0
zope.interface==4.5.0
